Fix Physics

+++ OMGBallz/OMGBallz/Util.cs
@@ -6,12 +6,12 @@ using System.Threading.Tasks;
 
 public static class Util
 {
-    public static float DistanceSquared(PhysicsObject first, PhysicsObject second)
+    public static double DistanceSquared(PhysicsObject first, PhysicsObject second)
     {
         return DistanceSquared(first.Position, second.Position);
     }
 
-    public static float DistanceSquared(Vector first, Vector second)
+    public static double DistanceSquared(Vector first, Vector second)
     {
         return (first - second).LengthSquared;
     }
